Remove cyclic import based on I/O in pdb.py.
Partially addresses https://github.com/jensengroup/propka-3.1/issues/49
